backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Candor Dust Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "argon2-cffi>=25.1.0",
    "python-multipart>=0.0.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "httpx>=0.27.0",
]

